use one cli file
